+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/test/java/com/liferay/dynamic/data/mapping/form/field/type/internal/date/time/DateTimeDDMFormFieldTypeSettingsTest.java
@@ -5,11 +5,11 @@
 
 package com.liferay.dynamic.data.mapping.form.field.type.internal.date.time;
 
-import com.liferay.dynamic.data.mapping.form.field.type.BaseDDMFormFieldTypeSettingsTestCase;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.model.DDMFormField;
 import com.liferay.dynamic.data.mapping.model.DDMFormLayout;
 import com.liferay.dynamic.data.mapping.model.DDMFormRule;
+import com.liferay.dynamic.data.mapping.test.util.BaseSettingsTestCase;
 import com.liferay.dynamic.data.mapping.test.util.DDMFormLayoutTestUtil;
 import com.liferay.dynamic.data.mapping.util.DDMFormFactory;
 import com.liferay.dynamic.data.mapping.util.DDMFormLayoutFactory;
@@ -26,8 +26,7 @@ import org.junit.Test;
 /**
  * @author Rodrigo Paulino
  */
-public class DateTimeDDMFormFieldTypeSettingsTest
-	extends BaseDDMFormFieldTypeSettingsTestCase {
+public class DateTimeDDMFormFieldTypeSettingsTest extends BaseSettingsTestCase {
 
 	@ClassRule
 	@Rule
