+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/test/java/com/liferay/dynamic/data/mapping/form/field/type/internal/util/NumericDDMFormFieldTypeUtilTest.java
@@ -5,9 +5,9 @@
 
 package com.liferay.dynamic.data.mapping.form.field.type.internal.util;
 
-import com.liferay.dynamic.data.mapping.form.field.type.BaseDDMFormFieldTypeSettingsTestCase;
 import com.liferay.dynamic.data.mapping.model.DDMFormField;
 import com.liferay.dynamic.data.mapping.render.DDMFormFieldRenderingContext;
+import com.liferay.dynamic.data.mapping.test.util.BaseSettingsTestCase;
 import com.liferay.dynamic.data.mapping.test.util.DDMFormValuesTestUtil;
 import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.util.HashMapBuilder;
@@ -24,8 +24,7 @@ import org.junit.Test;
 /**
  * @author Carolina Barbosa
  */
-public class NumericDDMFormFieldTypeUtilTest
-	extends BaseDDMFormFieldTypeSettingsTestCase {
+public class NumericDDMFormFieldTypeUtilTest extends BaseSettingsTestCase {
 
 	@ClassRule
 	@Rule
